+++ packages/sdks/javascript/src/services/SubnetService.ts
@@ -1,4 +1,3 @@
 import { SubnetsApiService } from "../api/generated/apis/subnets-api";
 
-export class SubnetService extends SubnetsApiService {
-}
+export class SubnetService extends SubnetsApiService {}