+++ packages/ui/react/src/components/Primitives/Dropdown/Dropdown.styled.ts
@@ -14,12 +14,10 @@ type StyledProps = {
    *
    * @see {@link https://styled-components.com/docs/api#transient-props | Styled Components - Transient props}
    */
-  [
-    Key in keyof Pick<
-      DropdownOption,
-      'activeBackgroundColor' | 'brokenSequence' | 'isActive' | 'color'
-    > as `$${Key}`
-  ]: DropdownOption[Key]
+  [Key in keyof Pick<
+    DropdownOption,
+    'activeBackgroundColor' | 'brokenSequence' | 'isActive' | 'color'
+  > as `$${Key}`]: DropdownOption[Key]
 } & {
   disabled: DropdownOption['isDisabled']
 }
@@ -101,12 +99,10 @@ type ContentBadgeProps = {
    *
    * @see {@link https://styled-components.com/docs/api#transient-props | Styled Components - Transient props}
    */
-  [
-    Key in keyof Pick<
-      NonNullable<DropdownOption['badge']>,
-      'backgroundColor'
-    > as `$${Key}`
-  ]: NonNullable<DropdownOption['badge']>[Key]
+  [Key in keyof Pick<
+    NonNullable<DropdownOption['badge']>,
+    'backgroundColor'
+  > as `$${Key}`]: NonNullable<DropdownOption['badge']>[Key]
 }
 
 const ContentBadge = styled.span<Pick<ContentBadgeProps, '$backgroundColor'>>`