+++ packages/ui/react/src/components/Primitives/Dropdown/Dropdown.tsx
@@ -206,7 +206,8 @@ type DropdownVariationContainerProps = DropdownSharedProps & {
 }
 
 type DropdownProps =
-  DropdownVariationContainerProps | DropdownVariationButtonProps
+  | DropdownVariationContainerProps
+  | DropdownVariationButtonProps
 
 type ElementAnchorProps = {
   popupState: PopupStateProps